# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BAM Store API",
        "description": "API Documentation for BAM Store",
        "contact": {
            "email": "admin@bamstore.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/admin/vendors": {
            "get": {
                "tags": [
                    "Admin Vendors"
                ],
                "summary": "Get list of vendors",
                "operationId": "f13782afbcbbd657ec0dc7b9f158e72a",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Vendors"
                ],
                "summary": "Create a new vendor",
                "operationId": "842bfdaf1b2a0922244cecc6cdd72774",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Vendor A"
                                    },
                                    "contact_info": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vendor created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/vendors/{id}": {
            "get": {
                "tags": [
                    "Admin Vendors"
                ],
                "summary": "Get vendor details",
                "operationId": "f67a4846fdb290a23492de58bc7ed6c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vendor ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Vendors"
                ],
                "summary": "Update a vendor",
                "operationId": "96e680d766185a772369b6650a24bfdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vendor ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "contact_info": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vendor updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Vendors"
                ],
                "summary": "Delete a vendor",
                "operationId": "900e79988b6230e9f3cd93ede55b1407",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vendor ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vendor deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/vendor-payments": {
            "get": {
                "tags": [
                    "Admin Vendor Payments"
                ],
                "summary": "Get list of vendor payments",
                "operationId": "4ffb0ab56b284102cbe9be8062c93bc7",
                "parameters": [
                    {
                        "name": "vendor_id",
                        "in": "query",
                        "description": "Filter by Vendor ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Vendor Payments"
                ],
                "summary": "Create a new vendor payment",
                "operationId": "5b4ec01335b54052986d6a8c2e7f2d39",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendor_id",
                                    "amount",
                                    "type",
                                    "payment_date"
                                ],
                                "properties": {
                                    "vendor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50000
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "dp",
                                            "installment",
                                            "full_payment"
                                        ],
                                        "example": "dp"
                                    },
                                    "payment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-01"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Down payment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/vendor-payments/{id}": {
            "delete": {
                "tags": [
                    "Admin Vendor Payments"
                ],
                "summary": "Delete a vendor payment",
                "operationId": "b7ba1ad3edc8aecb8162c1e06da1388b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Payment deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/members/search": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Search members by name or phone",
                "operationId": "28bb2eb455cf1ee6b925038c43387945",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query (min 3 chars)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product details",
                "operationId": "c5ec5b4fb4c3790c73e9849244665f45",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}